+++ spec/rspec/core/configuration_spec.rb
@@ -1293,17 +1293,6 @@ module RSpec::Core
 
     end
 
-    describe "#run_all_when_everything_filtered?" do
-      it "defaults to false" do
-        expect(config.run_all_when_everything_filtered?).to be(false)
-      end
-
-      it "can be queried by predicate method" do
-        config.run_all_when_everything_filtered = true
-        expect(config.run_all_when_everything_filtered?).to be(true)
-      end
-    end
-
     describe "#color_mode" do
       context ":automatic" do
         before do